+++ src/components/pages/movie-page-details/movie-page-details.tsx
