+++ src/components/pages/movie-page-in-list/movie-page-in-list.tsx
