+++ src/components/pages/movie-page/movie-page.tsx
@@ -84,17 +84,17 @@ function MoviePage(): React.FunctionComponent {
               <nav className="film-nav film-card__nav">
                 <ul className="film-nav__list">
                   <li className="film-nav__item film-nav__item--active">
-                    <a href="src/components/pages/moviePage/moviePage#MoviePage.tsx" className="film-nav__link">
+                    <a href="src/components/pages/movie-page/movie-page#MoviePage.tsx" className="film-nav__link">
                       Overview
                     </a>
                   </li>
                   <li className="film-nav__item">
-                    <a href="src/components/pages/moviePage/moviePage#MoviePage.tsx" className="film-nav__link">
+                    <a href="src/components/pages/movie-page/movie-page#MoviePage.tsx" className="film-nav__link">
                       Details
                     </a>
                   </li>
                   <li className="film-nav__item">
-                    <a href="src/components/pages/moviePage/moviePage#MoviePage.tsx" className="film-nav__link">
+                    <a href="src/components/pages/movie-page/movie-page#MoviePage.tsx" className="film-nav__link">
                       Reviews
                     </a>
                   </li>
