+++ src/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import App from './components/app/App';
+import App from './components/app/app';
 
 const defaultProps = {
   movieName: 'The Grand Budapest Hotel',
